Dobble Game
Dobble is a game in which players have to find symbols in common between two cards. Dobble is a speed and observation game for the whole family. 
The aim of the game?
Each two cards have one symbol in common. Be the first to find and name it to win the card.