Tag Identification using NLP
This project is a simple web scraper that takes a website link as input and returns the top 30 most common words in it,
excluding stopwords. It uses the nltk, urllib, BeautifulSoup, html5lib and matplotlib libraries to perform the tasks of
tokenization, web request, html parsing, frequency distribution and visualization.