Smart Power Consumption Analysis and Classification for Electrical Appliances
Smart Power Consumption Analysis and Classification for Electrical Appliances in Homes and Industries Using SVM Algorithm. 
Machine Learning Classification Algorithm : Support Vector Machine(SVM)
Prediction of Power consumption for different Electrical Appliances in Home and industries is  risky or not.
In industries and homes there are lot of Electrical Appliances/Machines, but if we have a problem with certain electrical appliance, we don't know the power consumption of that electrical appliance. They monitors all electrical appliances in terms of units. But by using ML algorithm we can able to classify by analyzing different parameters( current, voltage, power) to make decision whether the electrical appliance is Risky, Medium, Normal, No-load.
The data is acquired with the help of Raspberry Pi (series of single-board computers) and current, voltage, power sensors